hirehop/requirements.txt
asgiref==3.6.0
backports.zoneinfo==0.2.1
beautifulsoup4==4.11.1
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
cryptography==39.0.0
Django==4.1.5
django-bootstrap-v5==1.0.11
django-bootstrap5==22.2
gunicorn==20.1.0
idna==3.4
lxml==4.9.2
ntlm-auth==1.5.0
psycopg2-binary==2.9.5
pycparser==2.21
PyYAML==6.0
requests==2.28.1
requests-ntlm==1.1.0
requests-toolbelt==0.10.1
SharePlum==0.5.1
soupsieve==2.3.2.post1
sqlparse==0.4.3
urllib3==1.26.14